+++ src/Plutarch/Extra/DebuggableScript.hs
@@ -1,12 +1,17 @@
-{-# LANGUAGE NamedFieldPuns #-}
-{-# LANGUAGE OverloadedRecordDot #-}
 {-# LANGUAGE RankNTypes #-}
 
 module Plutarch.Extra.DebuggableScript (
-  DebuggableScript (..),
+  -- * Type
+  DebuggableScript,
+
+  -- * Construction
   checkedCompileD,
-  applyDebuggableScript,
   mustCompileD,
+
+  -- * Use
+  applyScript,
+  applyDebuggableScript,
+  applyDebuggableArg,
   mustFinalEvalDebuggableScript,
   finalEvalDebuggableScript,
   mustEvalScript,
@@ -17,6 +22,8 @@ import Control.DeepSeq (NFData)
 import Data.Text (Text)
 import qualified Data.Text as T
 import qualified Data.Text as Text
+import Optics.Getter (A_Getter, to, view)
+import Optics.Label (LabelOptic (labelOptic))
 import Plutarch (
   Config (Config, tracingMode),
   TracingMode (DetTracing, NoTracing),
@@ -25,7 +32,16 @@ import Plutarch (
 import Plutarch.Evaluate (EvalError, evalScript)
 import Plutarch.Extra.Compile (mustCompile)
 import PlutusLedgerApi.V1 (Data, ExBudget, Script)
-import PlutusLedgerApi.V1.Scripts (applyArguments)
+import PlutusLedgerApi.V1.Scripts (Script (Script), applyArguments)
+import UntypedPlutusCore (
+  Program (
+    Program,
+    _progAnn,
+    _progTerm,
+    _progVer
+  ),
+ )
+import qualified UntypedPlutusCore.Core.Type as UplcType
 import UntypedPlutusCore.Evaluation.Machine.Cek (
   CekUserError (CekEvaluationFailure, CekOutOfExError),
   ErrorWithCause (ErrorWithCause),
@@ -34,14 +50,9 @@ import UntypedPlutusCore.Evaluation.Machine.Cek (
 
 {- | A 'Script' with a debug fallback that has tracing turned on.
 
- @since 3.0.2
+ @since 3.8.0
 -}
-data DebuggableScript = DebuggableScript
-  { script :: Script
-  -- ^ @since 3.0.2
-  , debugScript :: Script
-  -- ^ @since 3.0.2
-  }
+data DebuggableScript = DebuggableScript Script Script
   deriving stock
     ( -- | @since 3.0.2
       Eq
@@ -55,16 +66,61 @@ data DebuggableScript = DebuggableScript
       NFData
     )
 
+{- | Retrieves the non-debugging 'Script'. This is read-only, as allowing it to
+ change could break invariants.
+
+ @since 3.8.0
+-}
+instance LabelOptic "script" A_Getter DebuggableScript DebuggableScript Script Script where
+  labelOptic = to $ \(DebuggableScript x _) -> x
+
+{- | Retrieves the debugging 'Script'. This is read-only, as allowing it to
+ change could break invariants.
+
+ @since 3.8.0
+-}
+instance LabelOptic "debugScript" A_Getter DebuggableScript DebuggableScript Script Script where
+  labelOptic = to $ \(DebuggableScript _ x) -> x
+
+{- | Apply a function to an argument on the compiled 'Script' level.
+
+ @since 3.8.0
+-}
+applyScript :: Script -> Script -> Script
+applyScript f a =
+  if fVer /= aVer
+    then error "apply: Plutus Core version mismatch"
+    else
+      Script
+        Program
+          { _progTerm = UplcType.Apply () fTerm aTerm
+          , _progVer = fVer
+          , _progAnn = ()
+          }
+  where
+    (Script Program {_progTerm = fTerm, _progVer = fVer}) = f
+    (Script Program {_progTerm = aTerm, _progVer = aVer}) = a
+
 {- | Apply given arguments to 'DebuggableScript'.
 
  @since 3.7.1
 -}
 applyDebuggableScript :: DebuggableScript -> [Data] -> DebuggableScript
-applyDebuggableScript (DebuggableScript script debugScript) d =
+applyDebuggableScript (DebuggableScript script debugScript) args =
+  DebuggableScript (applyArguments script args) (applyArguments debugScript args)
+
+{- | Apply a single argument, provided as a 'DebuggableScript'.
+
+ @since 3.8.0
+-}
+applyDebuggableArg ::
+  DebuggableScript ->
+  DebuggableScript ->
   DebuggableScript
-    { script = applyArguments script d
-    , debugScript = applyArguments debugScript d
-    }
+applyDebuggableArg f x =
+  DebuggableScript
+    (applyScript (view #script f) (view #script x))
+    (applyScript (view #debugScript f) (view #debugScript x))
 
 {- | For handling compilation errors right away.
 
@@ -80,7 +136,7 @@ checkedCompileD ::
 checkedCompileD term = do
   script <- compile Config {tracingMode = NoTracing} term
   debugScript <- compile Config {tracingMode = DetTracing} term
-  pure $ DebuggableScript {script, debugScript}
+  pure $ DebuggableScript script debugScript
 
 -- Like 'mustCompile', but with tracing turned on.
 mustCompileTracing ::
@@ -108,10 +164,7 @@ mustCompileD ::
   (forall (s :: S). Term s a) ->
   DebuggableScript
 mustCompileD term =
-  DebuggableScript
-    { script = mustCompile term
-    , debugScript = mustCompileTracing term
-    }
+  DebuggableScript (mustCompile term) (mustCompileTracing term)
 
 {- | Final evaluation of a 'DebuggableScript' to a 'Script', with errors resulting in
  exceptions.
@@ -143,7 +196,7 @@ mustFinalEvalDebuggableScript s =
 finalEvalDebuggableScript ::
   DebuggableScript ->
   (Either EvalError Script, ExBudget, [Text])
-finalEvalDebuggableScript DebuggableScript {script, debugScript} =
+finalEvalDebuggableScript (DebuggableScript script debugScript) =
   case res of
     Right _ -> r
     Left (ErrorWithCause evalErr _) ->
@@ -240,6 +293,5 @@ mustEvalD :: DebuggableScript -> DebuggableScript
 --   back to, so we need only 'mustEvalScript'.
 mustEvalD ds =
   DebuggableScript
-    { script = mustFinalEvalDebuggableScript ds
-    , debugScript = mustEvalScript (ds.debugScript)
-    }
+    (mustFinalEvalDebuggableScript ds)
+    (mustEvalScript . view #debugScript $ ds)