+++ src/Plutarch/Orphans.hs
@@ -1,9 +1,7 @@
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE QuantifiedConstraints #-}
-{-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
+-- The whole point of this module
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 {- | Module: Plutarch.Orphans
@@ -112,7 +110,7 @@ newtype AsBase16Bytes (a :: Type) = AsBase16Bytes a
 {- | Represent any serializable value as a hex-encoded JSON String of its
  serialization
 -}
-newtype AsBase16Codec (a :: Type) = AsBase16Codec {unAsBase16Codec :: a}
+newtype AsBase16Codec (a :: Type) = AsBase16Codec a
 
 --------------------
 -- Instances for `deriving via`
@@ -146,12 +144,12 @@ instance
 
 -- @ since 3.6.1
 instance (Serialise a) => Aeson.ToJSON (AsBase16Codec a) where
-  toJSON =
+  toJSON (AsBase16Codec x) =
     Aeson.String
       . encodeByteString
       . toStrict
       . serialise @a
-      . unAsBase16Codec
+      $ x
 
 -- @ since 3.6.1
 instance (Serialise a) => Aeson.FromJSON (AsBase16Codec a) where