+++ test/old/Spec/Extra/Map.hs
@@ -1,6 +1,3 @@
-{-# LANGUAGE PartialTypeSignatures #-}
-{-# LANGUAGE TypeApplications #-}
-
 module Spec.Extra.Map (tests) where
 
 --------------------------------------------------------------------------------